vim-plugins: add Rust support
Using Ansible's stat module to determine if a package should be
installed because Cargo errors if it tries to install a package that is
already installed, and re-installing all pacakges (because of `--force`)
takes a while.

There is no equivalent 'update' task because Cargo does not currently
implement an easy update mechanism for global packages.  See
https://github.com/rust-lang/cargo/issues 2082 for more details.

See:

- https://github.com/rust-lang/rust.vim
- https://github.com/rust-lang-nursery/rustfmt